requirements.txt
langchain
langchain-community
langchain-openai
faiss-cpu
pymupdf
pandas
python-dotenv
pydantic
tabulate